Remove valgrind test suite and support from compiletest
